requirements.txt
starlette
jinja2
uvicorn
itsdangerous
python-dotenv
passlib
python-multipart
bcrypt
sqlalchemy